requirements-docs.txt
numpydoc
sphinx
sphinx-automodapi
sphinx-panels
sphinx_rtd_theme
